vcs: exit with $fatal for errors (#340)
When the simulation exits abnormally, we should let the user know
there're failures.

* Add the flag for completion of workload lists

We add a flag to indicate the completion of workload lists and let
the simulation exit normally at the end of the list.

Since this flag will be read only when simv_init fails, we expect
minor impact on the simulation performance.